[test-framework] Clean up a repository only if there's no snapshots in progress

+++ test/framework/src/main/java/org/elasticsearch/test/rest/ESRestTestCase.java
@@ -1033,7 +1033,7 @@ public abstract class ESRestTestCase extends ESTestCase {
                     adminClient().performRequest(new Request("DELETE", "/_snapshot/" + repoName + "/" + name));
                 }
             }
-            if (preserveReposUponCompletion() == false) {
+            if (preserveReposUponCompletion() == false && inProgressSnapshots.containsKey(repoName) == false) {
                 deleteRepository(repoName);
             }
         }
